+++ embedchain/config/apps/CustomAppConfig.py
@@ -1,20 +1,14 @@
 import logging
-from enum import Enum
 from typing import Any
 
 from chromadb.api.types import Documents, Embeddings
 from dotenv import load_dotenv
-from embedchain.apps.CustomApp import LlmModels
+
+from embedchain.models import EmbeddingFunctions, LlmModels
 
 from .BaseAppConfig import BaseAppConfig
 
 load_dotenv()
-
-
-class EmbeddingFunctions(Enum):
-    OPENAI = "OPENAI"
-    HUGGING_FACE = "HUGGING_FACE"
-    VERTEX_AI = "VERTEX_AI"
 
 
 class CustomAppConfig(BaseAppConfig):
@@ -22,7 +16,16 @@ class CustomAppConfig(BaseAppConfig):
     Config to initialize an embedchain custom `App` instance, with extra config options.
     """
 
-    def __init__(self, log_level=None, ef: EmbeddingFunctions = None, db=None, host=None, port=None, id=None, llm_model: LlmModels=None):
+    def __init__(
+        self,
+        log_level=None,
+        ef: EmbeddingFunctions = None,
+        db=None,
+        host=None,
+        port=None,
+        id=None,
+        llm_model: LlmModels = None,
+    ):
         """
         :param log_level: Optional. (String) Debug level
         ['DEBUG', 'INFO', 'WARNING', 'ERROR', 'CRITICAL'].
@@ -52,11 +55,11 @@ class CustomAppConfig(BaseAppConfig):
         """
         Langchains default function layout for embeddings.
         """
+
         def embed_function(texts: Documents) -> Embeddings:
             return embeddings.embed_documents(texts)
-        
-        return embed_function
 
+        return embed_function
 
     @staticmethod
     def embedding_function(embedding_function: EmbeddingFunctions):
@@ -67,15 +70,18 @@ class CustomAppConfig(BaseAppConfig):
 
         if embedding_function == EmbeddingFunctions.OPENAI:
             from langchain.embeddings import OpenAIEmbeddings
+
             embeddings = OpenAIEmbeddings()
             return CustomAppConfig.langchain_default_concept(embeddings)
-        
+
         elif embedding_function == EmbeddingFunctions.HUGGING_FACE:
             from langchain.embeddings import HuggingFaceEmbeddings
+
             embeddings = HuggingFaceEmbeddings()
             return CustomAppConfig.langchain_default_concept(embeddings)
 
         elif embedding_function == EmbeddingFunctions.VERTEX_AI:
             from langchain.embeddings import VertexAIEmbeddings
+
             embeddings = VertexAIEmbeddings()
             return CustomAppConfig.langchain_default_concept(embeddings)
