+++ embedchain/models/LlmModels.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class LlmModels(Enum):
+    OPENAI = "OPENAI"
+    ANTHROPHIC = "ANTHPROPIC"
+    VERTEX_AI = "VERTEX_AI"
